requirements.txt
matplotlib==3.8.1
numpy==1.26.2
Pillow==10.1.0
scikit_learn==1.3.2
scipy==1.11.4
tqdm==4.66.1
